Many to one logic moved to a separate file.
