+++ lib/aurora_uix/templates/basic/renderers/field_renderer.ex
@@ -13,7 +13,7 @@ defmodule Aurora.Uix.Web.Templates.Basic.Renderers.FieldRenderer do
   use Aurora.Uix.Web.CoreComponentsImporter
   import Aurora.Uix.Web.Templates.Basic.Helpers, only: [get_field: 3]
 
-  alias Aurora.Uix.Web.Templates.Basic.Renderer
+  alias Aurora.Uix.Web.Templates.Basic.Renderers.ManyToOne
   alias Aurora.Uix.Web.Templates.Basic.Renderers.OneToMany
 
   @doc """
@@ -76,39 +76,15 @@ defmodule Aurora.Uix.Web.Templates.Basic.Renderers.FieldRenderer do
   end
 
   @spec default_render(map()) :: Phoenix.LiveView.Rendered.t()
+  # Delegates one to many association rendering
   defp default_render(%{field: %{field_type: :one_to_many_association}} = assigns),
     do: OneToMany.render(assigns)
 
+  # Delegates one to many association rendering
+  defp default_render(%{field: %{field_type: :many_to_one_association}} = assigns),
+    do: ManyToOne.render(assigns)
+
   # Renders different field types with appropriate HTML structure and components
-
-  defp default_render(
-         %{
-           field: %{field_type: :many_to_one_association} = field_struct,
-           _auix: %{_path: %{name: field_name}} = auix
-         } = assigns
-       )
-       when is_atom(field_name) do
-    inner_elements = get_association_paths(field_struct, auix._configurations, :show)
-    association_label = get_in(auix._configurations, [field_struct.resource, :parsed_opts, :name])
-
-    assigns
-    |> put_in([:_auix, :_path], %{
-      tag: :group,
-      config: [group_id: "#{field_struct.html_id}", title: association_label],
-      inner_elements: inner_elements
-    })
-    |> put_in([:_auix, :_ignore_association_label], true)
-    |> Renderer.render()
-  end
-
-  defp default_render(%{field: %{field_type: :many_to_one_association}} = assigns) do
-    assigns
-    |> parse_many_to_one_value()
-    |> set_many_to_one_resource()
-    |> trim_path()
-    |> render()
-  end
-
   defp default_render(assigns) do
     input_classes =
       "block w-full rounded-md border-zinc-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
@@ -141,137 +117,6 @@ defmodule Aurora.Uix.Web.Templates.Basic.Renderers.FieldRenderer do
       </div>
     <% end %>
     """
-  end
-
-  @spec parse_many_to_one_value(map()) :: map()
-  defp parse_many_to_one_value(%{_auix: %{_path: %{name: name}}} = assigns) when is_atom(name),
-    do: assigns
-
-  defp parse_many_to_one_value(
-         %{_auix: %{_path: %{name: names}, _mode: :show}, auix_entity: entity} = assigns
-       )
-       when is_tuple(names) do
-    names
-    |> Tuple.to_list()
-    |> delete_last()
-    |> Enum.reduce(entity, &Map.get(&2, &1, %{}))
-    |> then(&Map.put(assigns, :auix_entity, &1))
-  end
-
-  defp parse_many_to_one_value(
-         %{_auix: %{_path: %{name: names}, _form: form, _mode: :form} = _auix} = assigns
-       )
-       when is_tuple(names) do
-    names
-    |> Tuple.to_list()
-    |> List.first()
-    |> then(&%{&1 => form[&1].value})
-    |> then(&Map.put(assigns, :auix_entity, &1))
-    |> put_in([:_auix, :_mode], :show)
-    |> parse_many_to_one_value()
-  end
-
-  @spec trim_path(map()) :: map()
-  defp trim_path(%{_auix: %{_path: %{name: name}}} = assigns) when is_atom(name), do: assigns
-
-  defp trim_path(%{_auix: %{_path: %{name: names}}} = assigns) do
-    names
-    |> Tuple.to_list()
-    |> List.last()
-    |> then(&put_in(assigns, [:_auix, :_path, :name], &1))
-  end
-
-  @spec set_many_to_one_resource(map()) :: map()
-  defp set_many_to_one_resource(%{_auix: %{_path: %{name: name}}} = assigns) when is_atom(name),
-    do: assigns
-
-  defp set_many_to_one_resource(
-         %{
-           _auix: %{_path: %{name: names}, _configurations: configurations} = auix,
-           field: parent_field_struct
-         } = assigns
-       ) do
-    ignore_association_label? = Map.get(auix, :_ignore_association_label, false)
-
-    field_struct =
-      configurations
-      |> get_in([parent_field_struct.resource, :parsed_opts, :name])
-      |> then(&Map.put(parent_field_struct, :label, &1))
-
-    field =
-      names
-      |> Tuple.delete_at(0)
-      |> Tuple.to_list()
-      |> Enum.reduce(field_struct, fn field_name, parent_field ->
-        parent_field
-        |> maybe_ignore_parent_label(ignore_association_label?)
-        |> get_many_to_one_field(field_name, configurations)
-      end)
-
-    assigns
-    |> put_in(
-      [
-        :_auix,
-        :_configurations,
-        field.resource,
-        :resource_config,
-        Access.key!(:fields),
-        field.field
-      ],
-      field
-    )
-    |> put_in([:_auix, :_resource_name], field.resource)
-  end
-
-  @spec get_many_to_one_field(map(), atom(), map()) :: map()
-  defp get_many_to_one_field(
-         %{data: %{resource: resource_name}} = parent_field,
-         field_name,
-         configurations
-       ) do
-    parent_field
-    |> Map.put(:resource, resource_name)
-    |> Map.delete(:data)
-    |> get_many_to_one_field(field_name, configurations)
-  end
-
-  defp get_many_to_one_field(
-         %{resource: resource_name, label: parent_label} = _parent_field,
-         field_name,
-         configurations
-       ) do
-    %{name: field_name}
-    |> get_field(configurations, resource_name)
-    |> struct(%{readonly: true, disabled: true})
-    |> Map.update(:label, "", &"#{parent_label}#{&1}")
-  end
-
-  @spec get_association_paths(map(), map(), atom()) :: list()
-  defp get_association_paths(field_struct, configurations, path_type) do
-    configurations
-    |> get_in([field_struct.data.resource, :defaulted_paths, path_type, :inner_elements])
-    |> Kernel.||([])
-    |> convert_to_many_to_one_paths(field_struct.field)
-  end
-
-  @spec convert_to_many_to_one_paths(list(), atom()) :: list()
-  defp convert_to_many_to_one_paths(paths, parent_field) do
-    Enum.map(paths, &convert_to_many_to_one_path(&1, parent_field))
-  end
-
-  @spec convert_to_many_to_one_path(map(), atom()) :: map()
-  defp convert_to_many_to_one_path(
-         %{tag: :field, name: field_name, inner_elements: inner_elements} = path,
-         parent_field
-       ) do
-    Map.merge(path, %{
-      name: {parent_field, field_name},
-      inner_elements: convert_to_many_to_one_paths(inner_elements, parent_field)
-    })
-  end
-
-  defp convert_to_many_to_one_path(%{inner_elements: inner_elements} = path, parent_field) do
-    Map.put(path, :inner_elements, convert_to_many_to_one_paths(inner_elements, parent_field))
   end
 
   # Returns select field options and multiple selection flag if applicable
@@ -345,20 +190,4 @@ defmodule Aurora.Uix.Web.Templates.Basic.Renderers.FieldRenderer do
   defp get_many_to_one_select_option(%{field: %{data: %{related_key: related_key}}}, entity) do
     {entity |> Map.get(related_key) |> to_string(), Map.get(entity, related_key)}
   end
-
-  @spec delete_last(list()) :: list()
-  defp delete_last([]), do: []
-
-  defp delete_last(list) do
-    list
-    |> Enum.reverse()
-    |> then(fn [_first | rest] -> rest end)
-    |> Enum.reverse()
-  end
-
-  @spec maybe_ignore_parent_label(map(), boolean()) :: map()
-  defp maybe_ignore_parent_label(%{label: label} = parent_field, false),
-    do: Map.put(parent_field, :label, "#{label} ")
-
-  defp maybe_ignore_parent_label(parent_field, true), do: Map.put(parent_field, :label, "")
 end
